refactor: rename tasks-board provider to context
Renamed `provider` to `context` in tasks-board-list-item components for consistency with naming conventions. Updated all related imports, exports, and file paths to reflect this change.
