+++ apps/nextjs/src/components/features/tasks/tasks-board/tasks-board-list-item/provider/index.ts
@@ -1,3 +1,3 @@
+export * from './context';
 export { useTasksBoardListItemInputContext } from './input-context';
 export { useTasksBoardListItemContext } from './list-item-context';
-export * from './provider';
